feat

+++ src/main/java/hanghae7e6/prototype/recruitpost/RecruitPostRepositoryCustom.java
@@ -115,7 +115,7 @@ public class RecruitPostRepositoryCustom {
 
         Long postCount = query.fetchOne();
 
-        return postCount <= (dto.getOffSet() + 1) * dto.getLimit();
+        return postCount <= (long) (dto.getOffSet() + 1) * dto.getLimit();
     }
